Rename component to RecentlyPublishedDashlet - #4763
